+++ examples/threaded_progress_bar.py
@@ -5,7 +5,7 @@
 
 from __future__ import print_function
 
-__date__ = '2014-05-21'  # YYYY-MM-DD
+__date__ = '2014-06-16'  # YYYY-MM-DD
 
 try:
     import Queue as queue  # Python 2
@@ -38,15 +38,15 @@ class WorkerThread(threading.Thread):
 
 if __name__ == '__main__':
     # Create a message queue for communication in the main thread
-    msgqueue = queue.queue()
+    msgqueue = queue.Queue()
     worker_count = random.randint(3, 10)
 
     # Set the total amount of work for the ProgressBar
     total_work = worker_count * 10
 
     bar = progress.ProgressBar("Searching database: [{progress}] "
-                               "{percentage: .f}",
-                               mx=total_work)
+                               "{percentage: .0f}%",
+                               max=total_work)
 
     print("Creating {} worker threads...".format(worker_count))
 
@@ -54,16 +54,26 @@ if __name__ == '__main__':
     workers = [WorkerThread(msgqueue) for _ in range(worker_count)]
 
     # Start all workers
-    map(lambda x: x.start(), workers)
+    for w in workers:
+        w.start()
+
+    # Display the progress bar
+    bar.show()
 
     # Continue updating the progress bar while all workers are alive
-    while any(lambda x: x.is_alive(), workers):
+    while any(w.is_alive() for w in workers):
         try:
             bar.update(msgqueue.get_nowait())
+            bar.show()
         except queue.Empty:
             pass
 
+    # Show completion and print a newline
+    bar.show()
+    print()
+
     # Join all workers
-    map(lambda x: x.join(), workers)
+    for w in workers:
+        w.join()
 
     print("All worker threads finished...")